moved UpdateRedemptionRates and UpdateRedemptionRateForHostZone
